Tests and creation of receipt in service modified (not done yet)

+++ Evaluacion_1_Tingeso/src/main/java/com/example/Evaluacion_1_Tingeso/entities/ReceiptEntity.java
@@ -26,7 +26,7 @@ public class ReceiptEntity {
     private LocalTime workshopInHour;
 
     @Column(name = "typeOfRepairId")
-    private int typeOfRepairId;
+    private Long typeOfRepairId;
 
     @Column(name = "totalAmount")
     private int totalAmount;
